Refactor user management routes and add navigation HTML
This commit refactors the user management routes in the index.js file. It introduces a helper function called getNavigation() to generate the navigation HTML for the application. The function is used to display navigation links in the root route ("/") and the users route ("/users"). Additionally, the commit adds a link to view all users in the root route ("/") and a link to edit and delete individual users in the users route ("/users").